+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [2.3.2] - 2024-05-21
+### Added
+- Token, budget and tactic columns to the fantasy standings
+
 ## [2.3.1] - 2024-05-17
 ### Added
 - Navigator menus